Updated main.cpp and hpp

+++ pozdeev.andrey/P3/matrix_utils.hpp
@@ -5,7 +5,7 @@
 #include <fstream>
 
 namespace pozdeev {
-  constexpr int MAX_STATIC_SIZE = 10000;
+  constexpr size_t MAX_STATIC_SIZE = 10000;
 
   int countNonSequentialRows(const int * matrix, size_t rows, size_t cols);
   void spiral(int * matrix, size_t rows, size_t cols);
